tests/wpt/tests/fullscreen/model/remove-parent.html

<!DOCTYPE html>
<title>Remove the parent of the fullscreen element</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<div>
  <div id="child"></div>
</div>
<script>
  promise_test(async (t) => {
    t.add_cleanup(() => {
      if (document.fullscreenElement) {
        return document.exitFullscreen();
      }
    });
    const child = document.getElementById("child");
    const [, event] = await Promise.all([
      trusted_request(child),
      fullScreenChange(),
    ]);
    assert_equals(document.fullscreenElement, child);
    assert_equals(event.target, child);
    child.parentNode.remove();
    // Because /child/ was removed from the top layer, the fullscreen
    // element becomes null synchronously.
    assert_equals(document.fullscreenElement, null);
    const secondEvent = await fullScreenChange();
    assert_equals(document.fullscreenElement, null);
    assert_equals(secondEvent.target, document);
  });
</script>
