LayoutTests/imported/w3c/web-platform-tests/fullscreen/model/remove-parent-expected.txt


FAIL Remove the parent of the fullscreen element assert_equals: expected Document node with 2 children but got Element node <html><head><title>Remove the parent of the fullscreen el...
